Update NWN.Core. Split callbacks to new class for easier setup.
