+++ README.md
@@ -1,5 +1,5 @@
 # NWN .NET Barebone Template #
-This is a minimal (one file) implementation of NWNXEE DotNet using the [NWN.Core](https://github.com/nwn-dotnet/NWN.Core) library. It is intended for those who would like to build their own implementations from the ground up. See documentation in comments of Main.cs.
+This is a minimal implementation of NWNXEE DotNet using the [NWN.Core](https://github.com/nwn-dotnet/NWN.Core) library. It is intended for those who would like to build their own implementations from the ground up. See documentation in comments of Main.cs.
 
 ### Credits ###
 * Barebone implementation by [Wytchwood](https://github.com/Wytchwood)
